requirements.txt
langchain[llms]
openai
python-dotenv
GitPython